refactored dapp figma
