+++ apps/remix-ide/src/app/plugins/remixAI/index.ts
@@ -8,3 +8,5 @@ export { ModelManager } from './ModelManager'
 export type { ModelManagerDeps } from './ModelManager'
 export { DeepAgentManager } from './DeepAgentManager'
 export type { DeepAgentManagerDeps } from './DeepAgentManager'
+export { DAppGenerationManager } from './DAppGenerationManager'
+export type { DAppGenerationManagerDeps } from './DAppGenerationManager'
