+++ apps/remix-ide/src/app/plugins/remixAIPlugin.tsx
@@ -9,7 +9,7 @@ import { RemixMCPServer, createRemixMCPServer } from '@remix/remix-ai-core';
 import { AIModel, getDefaultModel } from '@remix/remix-ai-core';
 import axios from 'axios';
 import { endpointUrls } from "@remix-endpoints-helper"
-import { DeepAgentEventBridge, MCPServerManager, PermissionChecker, ModelManager, DeepAgentManager, ChatRequestBuffer } from './remixAI'
+import { DeepAgentEventBridge, MCPServerManager, PermissionChecker, ModelManager, DeepAgentManager, DAppGenerationManager, ChatRequestBuffer } from './remixAI'
 
 const profile = {
   name: 'remixAI',
@@ -53,10 +53,10 @@ export class RemixAIPlugin extends Plugin {
   remoteInferencer:RemoteInferencer | OllamaInferencer | MCPInferencer = null
   isInferencing: boolean = false
   chatRequestBuffer: ChatRequestBuffer<any> = null
-  codeExpAgent: CodeExplainAgent
-  securityAgent: SecurityAgent
-  contractor: ContractAgent
-  workspaceAgent: workspaceAgent
+  codeExpAgent: CodeExplainAgent | null = null
+  securityAgent: SecurityAgent | null = null
+  contractor: ContractAgent | null = null
+  workspaceAgent: workspaceAgent | null = null
   modelAccess: any
   selectedModel: AIModel = getDefaultModel() // default model
   selectedModelId: string = getDefaultModel().id
@@ -77,12 +77,15 @@ export class RemixAIPlugin extends Plugin {
   private permissionChecker: PermissionChecker
   private modelManager: ModelManager
   private deepAgentManager: DeepAgentManager
+  private dappManager: DAppGenerationManager
 
   constructor() {
     super(profile)
     this.eventBridge = new DeepAgentEventBridge()
     this.mcpManager = new MCPServerManager(this as any)
     this.permissionChecker = new PermissionChecker()
+    this.deepAgentEnabled = true
+
     this.modelManager = new ModelManager({
       plugin: this as any,
       eventBridge: this.eventBridge,
@@ -94,6 +97,7 @@ export class RemixAIPlugin extends Plugin {
       mcpManager: this.mcpManager,
       setupDeepAgentEventListeners: () => this.setupDeepAgentEventListeners()
     })
+    this.dappManager = new DAppGenerationManager({ plugin: this as any })
     // Set up MCP manager deps after all managers are created
     this.mcpManager.setDeps({
       plugin: this as any,
@@ -183,11 +187,7 @@ export class RemixAIPlugin extends Plugin {
     const allTools = await this.mcpInferencer?.getAllTools();
     console.log('[RemixAI Plugin] MCP tools available after wait:', allTools);
 
-    // Initialize DeepAgent if enabled (API key handled by proxy server)
-    const deepAgentEnabled = true // localStorage.getItem('deepagent_enabled') === 'true'
-    console.log('[RemixAI Plugin] DeepAgent enabled in localStorage:', deepAgentEnabled);
-
-    if (deepAgentEnabled && this.remixMCPServer) {
+    if (this.deepAgentEnabled && this.remixMCPServer) {
       try {
         console.log('[RemixAI Plugin] Initializing DeepAgent with mcpInferencer:', !!this.mcpInferencer);
         console.log('[RemixAI Plugin] Using model for DeepAgent:', this.selectedModel.provider, this.selectedModelId);
@@ -204,8 +204,6 @@ export class RemixAIPlugin extends Plugin {
           { provider: this.selectedModel.provider as 'anthropic' | 'mistralai', modelId: this.selectedModelId } // Pass selected model
         )
         await this.deepAgentInferencer.initialize()
-        this.deepAgentEnabled = true
-
         // Set up DeepAgent event listeners for streaming (once only)
         this.setupDeepAgentEventListeners();
 
@@ -657,67 +655,7 @@ export class RemixAIPlugin extends Plugin {
     isUpdate?: boolean;
     hasFigma?: boolean;
   }): Promise<string> {
-    try {
-      if (!this.deepAgentEnabled || !this.deepAgentInferencer) {
-        throw new Error('DeepAgent not enabled')
-      }
-
-      console.log('[QuickDapp] generateDAppContent called', {
-        messageCount: params.messages.length,
-        hasImage: params.hasImage,
-        isUpdate: params.isUpdate,
-        hasFigma: params.hasFigma
-      })
-
-      // Extract the user message content from messages
-      const lastUserMessage = params.messages[params.messages.length - 1]
-      let userContent = ''
-      let imageBase64: string | undefined
-
-      if (lastUserMessage) {
-        if (typeof lastUserMessage.content === 'string') {
-          userContent = lastUserMessage.content
-        } else if (Array.isArray(lastUserMessage.content)) {
-          // Handle multimodal content (text + image)
-          for (const part of lastUserMessage.content) {
-            if (part.type === 'text') {
-              userContent += part.text
-            } else if (part.type === 'image_url' && part.image_url?.url) {
-              imageBase64 = part.image_url.url
-            }
-          }
-        }
-      }
-
-      // Call DeepAgent with DApp Generator context
-      const generationParams: IParams = {
-        ...GenerationParams,
-        stream: false,
-        stream_result: false,
-        return_stream_response: false,
-        provider: 'anthropic',
-        model: 'claude-sonnet-4-5'
-      }
-
-      // Use DeepAgent answer method with the custom system prompt
-      const result = await this.deepAgentInferencer.answerWithCustomSystemPrompt(
-        userContent,
-        params.systemPrompt,
-        generationParams,
-        imageBase64
-      )
-
-      if (!result) {
-        throw new Error('No response from DeepAgent')
-      }
-
-      console.log('[QuickDapp] generateDAppContent result length:', result.length)
-      return result
-
-    } catch (error: any) {
-      console.error('[QuickDapp] generateDAppContent error:', error)
-      throw error // Re-throw so callers can handle the error
-    }
+    return this.dappManager.generateDAppContent(params)
   }
 
   /**
@@ -728,110 +666,7 @@ export class RemixAIPlugin extends Plugin {
     figmaUrl: string;
     figmaToken: string;
   }): Promise<{ success: boolean; fileName?: string; rawJson?: string; fileKey?: string; message?: string }> {
-    try {
-      // Parse Figma URL to extract file key
-      const patterns = [
-        /figma\.com\/(?:file|design)\/([a-zA-Z0-9]+)/,
-        /figma\.com\/proto\/([a-zA-Z0-9]+)/
-      ]
-
-      let fileKey: string | null = null
-      for (const pattern of patterns) {
-        const match = params.figmaUrl.match(pattern)
-        if (match) {
-          fileKey = match[1]
-          break
-        }
-      }
-
-      if (!fileKey) {
-        return { success: false, message: 'Invalid Figma URL format' }
-      }
-
-      // Fetch from Figma API
-      const apiUrl = `https://api.figma.com/v1/files/${fileKey}`
-      const response = await fetch(apiUrl, {
-        headers: {
-          'X-Figma-Token': params.figmaToken
-        }
-      })
-
-      if (!response.ok) {
-        if (response.status === 403) {
-          return { success: false, message: 'Figma API access denied. Please check your Personal Access Token.' }
-        }
-        if (response.status === 404) {
-          return { success: false, message: 'Figma file not found. Please check the URL.' }
-        }
-        return { success: false, message: `Figma API error: ${response.statusText}` }
-      }
-
-      const figmaData = await response.json()
-
-      // Extract relevant design data
-      const simplifyNode = (node: any, depth: number = 0): any => {
-        if (depth > 5) return null
-
-        const simplified: any = {
-          name: node.name,
-          type: node.type
-        }
-
-        if (node.absoluteBoundingBox) {
-          simplified.bounds = {
-            w: Math.round(node.absoluteBoundingBox.width),
-            h: Math.round(node.absoluteBoundingBox.height)
-          }
-        }
-
-        if (node.fills && node.fills.length > 0) {
-          const solidFill = node.fills.find((f: any) => f.type === 'SOLID' && f.visible !== false)
-          if (solidFill?.color) {
-            const r = Math.round((solidFill.color.r || 0) * 255)
-            const g = Math.round((solidFill.color.g || 0) * 255)
-            const b = Math.round((solidFill.color.b || 0) * 255)
-            simplified.fill = `#${r.toString(16).padStart(2, '0')}${g.toString(16).padStart(2, '0')}${b.toString(16).padStart(2, '0')}`
-          }
-        }
-
-        if (node.type === 'TEXT' && node.characters) {
-          simplified.text = node.characters.substring(0, 100)
-        }
-
-        if (node.children && Array.isArray(node.children)) {
-          const simplifiedChildren = node.children
-            .map((child: any) => simplifyNode(child, depth + 1))
-            .filter(Boolean)
-
-          if (simplifiedChildren.length > 0) {
-            simplified.children = simplifiedChildren
-          }
-        }
-
-        return simplified
-      }
-
-      const simplifiedDocument = simplifyNode(figmaData.document)
-      const rawJson = JSON.stringify(simplifiedDocument, null, 2)
-
-      // Truncate if too large
-      const maxJsonLength = 100000
-      const truncatedJson = rawJson.length > maxJsonLength
-        ? rawJson.substring(0, maxJsonLength) + '\n... [truncated for token limit]'
-        : rawJson
-
-      console.log('[QuickDapp] fetchFigmaDesign success:', figmaData.name)
-
-      return {
-        success: true,
-        fileKey,
-        fileName: figmaData.name || 'Untitled',
-        rawJson: truncatedJson
-      }
-    } catch (error: any) {
-      console.error('[QuickDapp] fetchFigmaDesign error:', error)
-      return { success: false, message: error.message || 'Failed to fetch Figma design' }
-    }
+    return this.dappManager.fetchFigmaDesign(params)
   }
 
   /**
@@ -847,37 +682,7 @@ export class RemixAIPlugin extends Plugin {
     systemPrompt: string;
     isBaseMiniApp?: boolean;
   }): Promise<string> {
-    try {
-      // First fetch the Figma design
-      const figmaResult = await this.fetchFigmaDesign({
-        figmaUrl: params.figmaUrl,
-        figmaToken: params.figmaToken
-      })
-
-      if (!figmaResult.success) {
-        throw new Error(figmaResult.message || 'Failed to fetch Figma design')
-      }
-
-      // Build description with Figma context
-      const enrichedDescription = `
-${params.description || 'Implement the design exactly as shown in the Figma file.'}
-
-**FIGMA DESIGN DATA:**
-File: ${figmaResult.fileName}
-${figmaResult.rawJson}
-`
-
-      // Generate DApp with Figma data
-      return await this.generateDAppContent({
-        messages: [{ role: 'user', content: enrichedDescription }],
-        systemPrompt: params.systemPrompt,
-        hasImage: false,
-        hasFigma: true
-      })
-    } catch (error: any) {
-      console.error('[QuickDapp] generateDAppFromFigma error:', error)
-      throw error
-    }
+    return this.dappManager.generateDAppFromFigma(params)
   }
 
   private async refreshMCPServersOnAuthChange(authState: any): Promise<void> {
